// api/src/__generated__/committee_members.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: z0sJvJScWtSrnKi4kVi4S6zmbl/H+A65jWrQkQ/0tA05GPaNt6EUI1y5zfmBjmWH0CFfQyH58FdWPHapJ6K7XQ==
 */

// eslint:disable
// tslint:disable

interface CommitteeMembers {
  active_year_id: number
  created_at: (Date) | null
  description: (string) | null
  favourite_game: (string) | null
  favourite_sugar: (string) | null
  /**
   * @default nextval('committee_members_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'committee_members_id'}
  image: (string) | null
  name: string
  quote: (string) | null
  role: string
  updated_at: (Date) | null
}
export default CommitteeMembers;

interface CommitteeMembers_InsertParameters {
  active_year_id: number
  created_at?: (Date) | null
  description?: (string) | null
  favourite_game?: (string) | null
  favourite_sugar?: (string) | null
  /**
   * @default nextval('committee_members_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'committee_members_id'}
  image?: (string) | null
  name: string
  quote?: (string) | null
  role: string
  updated_at?: (Date) | null
}
export type {CommitteeMembers_InsertParameters}